feat: seed location and review images
- Fix background fallback images
